Pub File I/O
File transfers work! After much tinkering, I've written a handler that
properly receives the map/pub files to the client machine upon request.
I have not tested this extensively but text diffs show the files are
identical (I tried this on a map file and eif file)

Additionally, I've written load/save methods for the 4 types of pub
files in existence. This is not extensively tested either (especially
save). The debugger showed correct values for npcs.