RID changes when saving pub files
The RID for each pub file is now updated to a CRC32 of the pub data
(excluding the first 10 bytes of each file).

Also restructed the ItemFile to not use LayoutKind.Explicit, instead, it
uses shared private members that are exposed publicly under different
names to recreate the same functionality.