In-game chat
Added some baseline code to support in-game chatting. This is a WIP.
Currently, the news loaded from the server is correctly displayed and
switching between chat tabs works normally. There is a bug with my
scrollbar implementation that I need to fix still but it is being a pain
in the ass so I'm ignoring it for now.