Map Rendering; Code Analysis
Improvements to map rendering and fixes for some exceptions that were
being thrown. One was happening because no locking was being done around
the NPC collection. The other was happening because multiple NPCs would
be entering the GFXLoader.TextureFromResource method, and duplicate key
errors would occur in the cache.

Also ran code analysis on the solution.