Hiding
Hidden characters drawn at 50% alpha when the main player is an admin, or not drawn when the main player is not an admin. Because of the way I do blending for the main character, the map will get slightly "brighter" if the main player types "$hide". Not sure of a way around this yet.