Minimap Rendering
The minimap is toggled on/off like the normal game. The minimap isn't perfect but it is more than good enough for the short amount of time it took to code up.

Also, added some changes to the EOChatBubble regarding the disposable pattern. Trying to debug why there are so many ObjectDisposedExceptions on the call to SpriteBatch.End() in it's draw method. Mostly will redesign how it is used and attach a single chat bubble to a single actor.