Chat Bubble improvements, Bug fixes, Stability
Chat bubbles are now a member variable that follows an actor around (NPC or character). This fixes the problems that the client was having with showing the same message for EVERY npc instead of the correct one, which I thought was a server issue but I guess not. I also modified the way that the EOGame removes UI components from the Components list so NPCs/chat bubbles aren't destroyed when switching game states. It makes sure the component being removed is actually a member of EOGame since those are the only things that should be removed. Game is so much more stable now.