Character States
Changed from boolean "Attacking" and "Walking" properties to a state-based "State" property. State can be Attacking, Walking, Sitting (in which case, SitState is used to determine floor vs chair), or Emote0-9. This paves the way for easily adding more actions to character. I'm not sure if I'll need a separate spell casting or if I can put that in attacking.