MapItems dictionary
With the intention of improving performance, I hid the MapItems collection from visibility and converted it to a dictionary. Item lookups by x/y coord are better now but I'm not sure about searching based on uid. Oh well. I also split out part of the update method to make it more readable.