Offset Adjustments
Adjustments to the offsets for certain things that show up above actors. The hover name (where the mouse cursor is) in the map renderer is more accurately centered for items, npcs, and characters. The health bar and damage counters for NPCs and players also show up a lot more nicely now. Added TopPixel to character renderer to help accomplish this.