Rendering Enhancement
The ground layer used to have a slight tearing problem that would occur when the character was walking (sometimes). This was due to the offsets being used to get draw coordinates for tiles were changing while the character walked, and sometimes this would happen in the middle of a draw call. This has been fixed. I also improved the synchronization for when a map change event is occurring.