Sleep between file requests
The file requests were being sent too fast if more than one file needed to be fetched from the server. This caused file transfers to fail, which would give a failed login message and not launch the game properly when starting from an empty installation. The thread sleeps disabled in debug builds have been enabled unconditionally again so that files are transferred properly.