Rendering additional fix
Based on results of profiling, it looked like the call to get the 'enable transition' value from the config was taking a lot of cycles. This is now loaded when the map renderer is created and stored in the map renderer. Also added lock around map drawing that is done in the update method for the render targets.