Item Equip Fixes
Two bugs fixed: one, equipping an item would sometimes leave a phantom copy in the inventory. This was due to checking for paperdoll dialog being null before handling the packet. Two, difficulty equipping items with a subloc (rings/armlets/bracers). This was due to not having an 'else' where there should have been one and essentially adding the item twice.