Working Sound
A few sounds have been added to the game. EOSoundManager has been rewritten a bit so that it works better with how it will be used. I had to add a function _correctTheFileLength so that certain files would load properly. I also fixed it such that multiple instances of the same sound effect can play simultaneously.