Trading
Trade check-in. Major changes in some files, mostly related to the trade dialog and trade API. Added some fixes and minor changes to other components. XNAControls is unchanged but I did a rebuild of the binary. The only bug is related to scrolling in a dialog window - still need to track why the events OnMouseEnter and OnMouseLeave are not firing for an XNA control, then all should be well.