Removed otherPlayers collection
otherPlayers (List) has been removed in favor of a single collection that manages character renderers. 'Character' is accessible via EOCharacterRenderer.Character anyway so it was really just a maintenance change. This also prevents enumeration errors when the collection changes while an enumerable operation is occurring on the list (which didn't have a locking mechanism around it). So, hopefully a stability fix too.