EOBot Refactor
Refactored EOBot to use an object-oriented approach. I don't like the way BotFramework is implemented so this will most likely change. However, IBot can be implemented (recommend deriving from BotBase) in order to implement a bot and use it within the framework. The framework needs to be more extensible to support IBots of varying types. Probably will use factory pattern to accomplish this.