Font Bug Fixed
Fixed the buggy font by refactoring XNAControls to use SpriteFont instead of System.Drawing.Font. There are a few remaining bugs that require investigating, but for the most part things are working OK. 1) display of labels that rely on auto-centering is a little blurry. 2) Underline for labels shows on top of text instead of underneath it. 3) There are intermittent screen flashes running on Windows 10 (not sure if debugging / OS related).