Suggested approach
Instead of injecting `IGameStateProvider` and comparing the current game state against `PlayingTheGame` for in-game handling, construct the character renderer with a value indicating if it is a UI control or not.

A more elegant solution would probably be deriving a separate type for character renderers that are used pre-game and don't have all the in-game garbage associated with them. I'll probably take care of this in the future as part of rendering code cleanup. For now this should be good enough.