Fix chest crash (#409)
Fixes crash bug due to dialog close operations resuming the rest of the Show() call in a different async thread context. This continues with disposal of the dialog, which removes the dialog from the game's components on a different thread, causing an exception if the update loop is in the process of iterating over the components elsewhere.